Rename url_constants.* to webui_url_constants
