+++ common/BUILD.gn
@@ -4,8 +4,8 @@ source_set("common") {
     "common_message_generator.h",
     "resource_bundle_helper.cc",
     "resource_bundle_helper.h",
-    "url_constants.cc",
-    "url_constants.h",
+    "webui_url_constants.cc",
+    "webui_url_constants.h",
   ]
   deps = [
     "extensions/api"
