+++ common/webui_url_constants.cc
@@ -1,4 +1,4 @@
-#include "brave/common/url_constants.h"
+#include "brave/common/webui_url_constants.h"
 
 const char kPaymentsHost[] = "payments";
 const char kPaymentsJS[] = "brave_payments.js";
